+++ src/main/java/com/exmachina/sam/currency/scheduledtasks/pojos/ThirdPartyRates.java
@@ -2,6 +2,7 @@ package com.exmachina.sam.currency.scheduledtasks.pojos;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 
+import java.math.BigDecimal;
 import java.util.LinkedHashMap;
 
-public class ThirdPartyRates extends LinkedHashMap<String , LinkedHashMap<String , Double>> {}
+public class ThirdPartyRates extends LinkedHashMap<String , LinkedHashMap<String , BigDecimal>> {}
